# docs/swagger/swagger.yaml
basePath: /api
definitions:
  github_com_alchemillahq_sylve_internal.APIResponse-any:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_alchemillahq_sylve_internal.APIResponse-array_github_com_alchemillahq_sylve_internal_db_models_zfs_PeriodicSnapshot:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_db_models_zfs.PeriodicSnapshot'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  internal_handlers_zfs.BulkDeleteRequest:
    properties:
      guids:
        items:
          type: string
        type: array
    required:
    - guids
    type: object
  internal_handlers_zfs.CreateFilesystemRequest:
    properties:
      name:
        type: string
      parent:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
    required:
    - name
    - parent
    type: object
  internal_handlers_zfs.CreatePeriodicSnapshotJobRequest:
    properties:
      cronExpr:
        type: string
      guid:
        type: string
      interval:
        type: integer
      prefix:
        type: string
      recursive:
        type: boolean
    required:
    - guid
    - interval
    - prefix
    type: object
  internal_handlers_zfs.CreateSnapshotRequest:
    properties:
      guid:
        type: string
      name:
        type: string
      recursive:
        type: boolean
    required:
    - guid
    - name
    type: object
...
  internal_handlers_zfs.DatasetListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.Dataset'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  internal_handlers_zfs.EditFilesystemRequest:
    properties:
      guid:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
    required:
    - guid
    - properties
    type: object
  internal_handlers_zfs.EditVolumeRequest:
    properties:
      name:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
    required:
    - name
    - properties
    type: object
...
  internal_handlers_zfs.RollbackSnapshotRequest:
    properties:
      destroyMoreRecent:
        type: boolean
      guid:
        type: string
    required:
    - guid
    type: object
...
host: sylve.lan:8181
info:
  contact:
    email: hello@alchemilla.io
    name: Alchemilla Ventures Pvt. Ltd.
    url: https://alchemilla.io
  description: Sylve is a lightweight GUI for managing Bhyve, Jails, ZFS, networking,
    and more on FreeBSD.
  license:
    name: BSD-2-Clause
    url: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  termsOfService: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  title: Sylve API
  version: 0.0.1
paths:
...
  /zfs/datasets:
    get:
      consumes:
      - application/json
      description: Get all ZFS datasets
      parameters:
      - description: Filter for datasets
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers_zfs.DatasetListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get all ZFS datasets
      tags:
      - ZFS
  /zfs/datasets/bulk-delete:
    post:
      consumes:
      - application/json
      description: Bulk delete ZFS datasets
      parameters:
      - description: Bulk Delete Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Bulk delete ZFS datasets
      tags:
      - ZFS
  /zfs/datasets/filesystem:
    patch:
      consumes:
      - application/json
      description: Edit a ZFS filesystem
      parameters:
      - description: Edit Filesystem Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.EditFilesystemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Edit a ZFS filesystem
      tags:
      - ZFS
    post:
      consumes:
      - application/json
      description: Create a ZFS filesystem
      parameters:
      - description: Create Filesystem Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.CreateFilesystemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Create a ZFS filesystem
      tags:
      - ZFS
  /zfs/datasets/filesystem/{guid}:
    delete:
      consumes:
      - application/json
      description: Delete a ZFS filesystem
      parameters:
      - description: Filesystem GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete a ZFS filesystem
      tags:
      - ZFS
  /zfs/datasets/snapshot:
    post:
      consumes:
      - application/json
      description: Create a ZFS snapshot
      parameters:
      - description: Create Snapshot Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.CreateSnapshotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Create a ZFS snapshot
      tags:
      - ZFS
  /zfs/datasets/snapshot/{guid}:
    delete:
      consumes:
      - application/json
      description: Delete a ZFS snapshot
      parameters:
      - description: Snapshot GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete a ZFS snapshot
      tags:
      - ZFS
  /zfs/datasets/snapshot/periodic:
    get:
      consumes:
      - application/json
      description: Get all periodic ZFS snapshots jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-array_github_com_alchemillahq_sylve_internal_db_models_zfs_PeriodicSnapshot'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get all periodic ZFS snapshot jobs
      tags:
      - ZFS
    post:
      consumes:
      - application/json
      description: Create a periodic ZFS snapshot job
      parameters:
      - description: Create Periodic Snapshot Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.CreatePeriodicSnapshotJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Create a periodic ZFS snapshot job
      tags:
      - ZFS
  /zfs/datasets/snapshot/periodic/{guid}:
    delete:
      consumes:
      - application/json
      description: Delete a periodic ZFS snapshot
      parameters:
      - description: Periodic Snapshot GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete a periodic ZFS snapshot
      tags:
      - ZFS
  /zfs/datasets/snapshot/rollback:
    post:
      consumes:
      - application/json
      description: Rollback to a ZFS snapshot
      parameters:
      - description: Rollback Snapshot Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.RollbackSnapshotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Rollback to a ZFS snapshot
      tags:
      - ZFS
  /zfs/datasets/volume:
    patch:
      consumes:
      - application/json
      description: Edit a ZFS volume
      parameters:
      - description: Edit Volume Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.EditVolumeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Edit a ZFS volume
      tags:
      - ZFS
    post:
      consumes:
      - application/json
      description: Create a ZFS volume
      parameters:
      - description: Create Volume Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.CreateVolumeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Create a ZFS volume
      tags:
      - ZFS
  /zfs/datasets/volume/{guid}:
    delete:
      consumes:
      - application/json
      description: Delete a ZFS volume
      parameters:
      - description: Volume GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete a ZFS volume
      tags:
      - ZFS
  /zfs/datasets/volume/flash:
    post:
      consumes:
      - application/json
      description: Flash a ZFS volume with a UUID pointing to a disk iso/img
      parameters:
      - description: Flash Volume Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.FlashVolumeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Flash a ZFS volume
      tags:
      - ZFS
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"